+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.1
+* support for videos with `youtube.be` address;
 * added the changelog file.